renamed some scripts #52
